[inga] Removed NODE_ID parameter from node-setup example
During the last commits the meaning of the NODE_ID was clarified.
It is used only provide a simple ID for testing and debugging purpose.
Not to be confused with the PAN_ADDR which is the short address of an 802.15.4 node.
The short address can be configured using the settings app.

Note that in general the EUI-64 is the most important identification number of each node.
The pan address is not required an can be derived from the EUI-64.